// src/app/Pages/Master Module/Models/HolidayMaster.ts
export class HolidayMaster {
    id: number;
    name: string = '';
    // TECHNICIAN_NAME: string = '';
    // LAB_NAME: string = '';
    date: any;
    is_active: boolean = true;
    reason: string = '';
    start_date;
    end_date;
    previous_start;
    previous_end;
    user_id: any;
    // HOSPITAL_ID: any;
    technician_id: any;
    lab_id: any;


}